+++ packages/vue/src/hooks/useMotion.ts
@@ -1,0 +1,1 @@
+export function useMotion() {}
