+++ packages/vue/src/index.ts
@@ -1,5 +1,5 @@
-export { default as Motion } from './component/Motion.vue'
-export { default as Presence } from './component/Presence.vue'
-export { default as PresenceGroup } from './component/PresenceGroup.vue'
+export { default as Motion } from './components/Motion.vue'
+export { default as Presence } from './components/Presence.vue'
+export { default as PresenceGroup } from './components/PresenceGroup.vue'
 
 export * from '@motionone/dom'
